a sad attempt to make the `env` initialization less of a mess
