+++ cider/src/flatten/structures/environment/env.rs
@@ -1,11 +1,10 @@
 use super::{
-    super::context::Context,
-    assignments::{AssignType, GroupInterfacePorts, ScheduledAssignments},
+    assignments::*,
     clock::{ClockMap, ReadSource},
-    program_counter::{
-        ControlTuple, ParEntry, PcMaps, ProgramCounter, WithEntry,
-    },
-    traverser::{Path, TraversalError},
+    maps::*,
+    program_counter::*,
+    traverser::{Path, TraversalError, Traverser},
+    wave::WaveWriter,
 };
 use crate::{
     configuration::{LoggingConfig, RuntimeConfig},
@@ -13,9 +12,7 @@ use crate::{
         self,
         commands::{ParseNodes, ParsePath},
     },
-    errors::{
-        BoxedCiderError, BoxedRuntimeError, CiderResult, ConflictingAssignments,
-    },
+    errors::*,
     flatten::{
         flat_ir::{
             base::{
@@ -26,352 +23,29 @@ use crate::{
             prelude::*,
             wires::guards::{Guard, PortComp},
         },
-        primitives::{
-            self, Primitive,
-            prim_trait::{RaceDetectionPrimitive, UpdateStatus},
-        },
+        primitives::{self, prim_trait::UpdateStatus},
         structures::{
-            context::{LookupName, PortDefinitionInfo},
-            environment::{
-                program_counter::ControlPoint, traverser::Traverser,
-            },
+            context::{Context, LookupName, PortDefinitionInfo},
             thread::{ThreadIdx, ThreadMap},
         },
         text_utils::Color,
     },
-    logging,
+    logging::*,
     serialization::{DataDump, MemoryDeclaration, PrintCode},
 };
-use crate::{
-    errors::{RuntimeError, RuntimeResult},
-    flatten::structures::environment::wave::WaveWriter,
+use std::{
+    collections::VecDeque,
+    convert::Into,
+    fmt::{Debug, Write},
 };
+
 use ahash::{HashMap, HashMapExt, HashSet, HashSetExt};
 use baa::{BitVecOps, BitVecValue};
 use calyx_frontend::source_info::PositionId;
-use cider_idx::{
-    IndexRef,
-    iter::IndexRange,
-    maps::{IndexedMap, SecondaryMap},
-};
+use cider_idx::{IndexRef, maps::SecondaryMap};
 use fxhash::{FxHashMap, FxHashSet};
 use itertools::Itertools;
 use owo_colors::OwoColorize;
-
-use slog::{Logger, info, warn};
-use std::{collections::VecDeque, fmt::Write};
-use std::{convert::Into, fmt::Debug};
-
-#[derive(Debug, Clone)]
-pub struct PortMap(IndexedMap<GlobalPortIdx, PortValue>);
-
-impl PortMap {
-    pub fn with_capacity(size: usize) -> Self {
-        Self(IndexedMap::with_capacity(size))
-    }
-}
-
-impl std::ops::DerefMut for PortMap {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-impl std::ops::Deref for PortMap {
-    type Target = IndexedMap<GlobalPortIdx, PortValue>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl PortMap {
-    /// Essentially asserts that the port given is undefined, it errors out if
-    /// the port is defined and otherwise does nothing
-    pub fn write_undef(&mut self, target: GlobalPortIdx) -> RuntimeResult<()> {
-        if self[target].is_def() {
-            Err(RuntimeError::UndefiningDefinedPort(target).into())
-        } else {
-            Ok(())
-        }
-    }
-
-    /// Sets the given index to the given value without checking whether or not
-    /// the assignment would conflict with an existing assignment. Should only
-    /// be used by cells to set values that may be undefined
-    pub fn write_exact_unchecked(
-        &mut self,
-        target: GlobalPortIdx,
-        val: PortValue,
-    ) -> UpdateStatus {
-        if self[target].is_undef() && val.is_undef()
-            || self[target]
-                .as_option()
-                .zip(val.as_option())
-                .map(|(a, b)| a.eq_no_transitive_clocks(b))
-                .unwrap_or_default()
-        {
-            UpdateStatus::Unchanged
-        } else {
-            self[target] = val;
-            UpdateStatus::Changed
-        }
-    }
-
-    #[inline]
-    pub(crate) fn insert_val_unchecked(
-        &mut self,
-        idx: GlobalPortIdx,
-        val: AssignedValue,
-    ) {
-        self[idx] = PortValue::new(val)
-    }
-
-    /// Sets the given index to undefined without checking whether or not it was
-    /// already defined
-    #[inline]
-    pub fn write_undef_unchecked(&mut self, target: GlobalPortIdx) {
-        self[target] = PortValue::new_undef();
-    }
-
-    #[inline(always)]
-    pub fn insert_val(
-        &mut self,
-        target: GlobalPortIdx,
-        val: AssignedValue,
-    ) -> Result<UpdateStatus, Box<ConflictingAssignments>> {
-        match self[target].as_option() {
-            Some(t) => {
-                if *t.winner() != AssignmentWinner::Implicit
-                    && t.has_conflict_with(&val)
-                {
-                    Err(ConflictingAssignments {
-                        target,
-                        a1: t.clone(),
-                        a2: val,
-                    }
-                    .into())
-                } else if t.eq_no_transitive_clocks(&val) {
-                    Ok(UpdateStatus::Unchanged)
-                } else {
-                    self[target] = PortValue::new(val);
-                    Ok(UpdateStatus::Changed)
-                }
-            }
-            // changed
-            None => {
-                self[target] = PortValue::new(val);
-                Ok(UpdateStatus::Changed)
-            }
-        }
-    }
-
-    /// Identical to `insert_val` but returns a `RuntimeError` instead of a
-    /// `ConflictingAssignments` error. This should be used inside of primitives
-    /// while the latter is used in the general simulation flow.
-    #[inline]
-    pub fn insert_val_general(
-        &mut self,
-        target: GlobalPortIdx,
-        val: AssignedValue,
-    ) -> RuntimeResult<UpdateStatus> {
-        self.insert_val(target, val)
-            .map_err(|e| RuntimeError::ConflictingAssignments(e).into())
-    }
-
-    pub fn set_done(
-        &mut self,
-        target: GlobalPortIdx,
-        done_bool: bool,
-    ) -> RuntimeResult<UpdateStatus> {
-        self.insert_val(
-            target,
-            AssignedValue::cell_value(if done_bool {
-                BitVecValue::new_true()
-            } else {
-                BitVecValue::new_false()
-            }),
-        )
-        .map_err(|e| RuntimeError::ConflictingAssignments(e).into())
-    }
-}
-
-pub(crate) type CellMap = IndexedMap<GlobalCellIdx, CellLedger>;
-pub(crate) type RefCellMap =
-    IndexedMap<GlobalRefCellIdx, Option<GlobalCellIdx>>;
-pub(crate) type RefPortMap =
-    IndexedMap<GlobalRefPortIdx, Option<GlobalPortIdx>>;
-pub(crate) type AssignmentRange = IndexRange<AssignmentIdx>;
-
-#[derive(Clone)]
-pub(crate) struct ComponentLedger {
-    pub(crate) index_bases: BaseIndices,
-    pub(crate) comp_id: ComponentIdx,
-}
-
-impl ComponentLedger {
-    /// Convert a relative offset to a global one. Perhaps should take an owned
-    /// value rather than a pointer
-    pub fn convert_to_global_port(&self, port: &PortRef) -> GlobalPortRef {
-        match port {
-            PortRef::Local(l) => (&self.index_bases + l).into(),
-            PortRef::Ref(r) => (&self.index_bases + r).into(),
-        }
-    }
-
-    pub fn convert_to_global_cell(&self, cell: &CellRef) -> GlobalCellRef {
-        match cell {
-            CellRef::Local(l) => (&self.index_bases + l).into(),
-            CellRef::Ref(r) => (&self.index_bases + r).into(),
-        }
-    }
-
-    pub fn signature_ports(&self, ctx: &Context) -> IndexRange<GlobalPortIdx> {
-        let sig = ctx.secondary[self.comp_id].signature();
-        let beginning = &self.index_bases + sig.start();
-        let end = &self.index_bases + sig.end();
-        IndexRange::new(beginning, end)
-    }
-}
-
-/// An enum encapsulating cell functionality. It is either a pointer to a
-/// primitive or information about a calyx component instance
-pub(crate) enum CellLedger {
-    Primitive {
-        // wish there was a better option with this one
-        cell_dyn: Box<dyn Primitive>,
-    },
-    RaceDetectionPrimitive {
-        cell_dyn: Box<dyn RaceDetectionPrimitive>,
-    },
-    Component(ComponentLedger),
-}
-
-impl Clone for CellLedger {
-    fn clone(&self) -> Self {
-        match self {
-            Self::Primitive { cell_dyn } => Self::Primitive {
-                cell_dyn: cell_dyn.clone_boxed(),
-            },
-            Self::RaceDetectionPrimitive { cell_dyn } => {
-                Self::RaceDetectionPrimitive {
-                    cell_dyn: cell_dyn.clone_boxed_rd(),
-                }
-            }
-            Self::Component(component_ledger) => {
-                Self::Component(component_ledger.clone())
-            }
-        }
-    }
-}
-
-impl From<ComponentLedger> for CellLedger {
-    fn from(v: ComponentLedger) -> Self {
-        Self::Component(v)
-    }
-}
-
-impl From<Box<dyn RaceDetectionPrimitive>> for CellLedger {
-    fn from(cell_dyn: Box<dyn RaceDetectionPrimitive>) -> Self {
-        Self::RaceDetectionPrimitive { cell_dyn }
-    }
-}
-
-impl From<Box<dyn Primitive>> for CellLedger {
-    fn from(cell_dyn: Box<dyn Primitive>) -> Self {
-        Self::Primitive { cell_dyn }
-    }
-}
-
-impl CellLedger {
-    fn new_comp<C: AsRef<Context> + Clone>(
-        idx: ComponentIdx,
-        env: &Environment<C>,
-    ) -> Self {
-        Self::Component(ComponentLedger {
-            index_bases: BaseIndices::new(
-                env.ports.peek_next_idx(),
-                (env.cells.peek_next_idx().index() + 1).into(),
-                env.ref_cells.peek_next_idx(),
-                env.ref_ports.peek_next_idx(),
-            ),
-            comp_id: idx,
-        })
-    }
-
-    pub fn as_comp(&self) -> Option<&ComponentLedger> {
-        match self {
-            Self::Component(comp) => Some(comp),
-            _ => None,
-        }
-    }
-
-    #[inline]
-    pub fn unwrap_comp(&self) -> &ComponentLedger {
-        self.as_comp()
-            .expect("Unwrapped cell ledger as component but received primitive")
-    }
-
-    #[must_use]
-    pub fn as_primitive(&self) -> Option<&dyn Primitive> {
-        match self {
-            Self::Primitive { cell_dyn } => Some(&**cell_dyn),
-            Self::RaceDetectionPrimitive { cell_dyn } => {
-                Some(cell_dyn.as_primitive())
-            }
-            _ => None,
-        }
-    }
-
-    pub fn unwrap_primitive(&self) -> &dyn Primitive {
-        self.as_primitive()
-            .expect("Unwrapped cell ledger as primitive but received component")
-    }
-}
-
-impl Debug for CellLedger {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Primitive { .. } => f.debug_struct("Primitive").finish(),
-            Self::RaceDetectionPrimitive { .. } => {
-                f.debug_struct("RaceDetectionPrimitive").finish()
-            }
-            Self::Component(ComponentLedger {
-                index_bases,
-                comp_id,
-            }) => f
-                .debug_struct("Component")
-                .field("index_bases", index_bases)
-                .field("comp_id", comp_id)
-                .finish(),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct PinnedPorts {
-    map: HashMap<GlobalPortIdx, BitVecValue>,
-}
-
-impl PinnedPorts {
-    pub fn iter(&self) -> impl Iterator<Item = (&GlobalPortIdx, &BitVecValue)> {
-        self.map.iter()
-    }
-
-    pub fn new() -> Self {
-        Self {
-            map: HashMap::new(),
-        }
-    }
-
-    pub fn insert(&mut self, port: GlobalPortIdx, val: BitVecValue) {
-        self.map.insert(port, val);
-    }
-
-    pub fn remove(&mut self, port: GlobalPortIdx) {
-        self.map.remove(&port);
-    }
-}
 
 #[derive(Debug, Clone)]
 pub struct Environment<C: AsRef<Context> + Clone> {
@@ -401,11 +75,271 @@ pub struct Environment<C: AsRef<Context> + Clone> {
     memory_header: Option<Vec<MemoryDeclaration>>,
     logger: Logger,
 
-    group_instances: Vec<(GlobalCellIdx, GroupIdx)>,
     ports_to_cells_map: SecondaryMap<GlobalPortIdx, GlobalCellIdx>,
 }
 
 impl<C: AsRef<Context> + Clone> Environment<C> {
+    pub fn new(
+        ctx: C,
+        data_map: Option<DataDump>,
+        check_race: bool,
+        logging_conf: LoggingConfig,
+    ) -> Self {
+        /// Internal function used to layout a given component from a cell id
+        ///
+        /// Layout is handled in the following order:
+        /// 1. component signature (input/output)
+        /// 2. group hole ports
+        /// 3. cells + ports, primitive
+        /// 4. sub-components
+        /// 5. ref-cells & ports
+        fn layout_component<C: AsRef<Context> + Clone>(
+            env: &mut Environment<C>,
+            ctx: &Context,
+            comp: GlobalCellIdx,
+            data_map: &Option<DataDump>,
+            memories_initialized: &mut HashSet<String>,
+            check_race: bool,
+        ) {
+            let ComponentLedger {
+                index_bases,
+                comp_id,
+            } = env.cells[comp]
+                .as_comp()
+                .expect("Called layout component with a non-component cell.");
+            let comp_aux = &ctx.secondary[*comp_id];
+
+            // Insert the component's continuous assignments into the program counter, if non-empty
+            let cont_assigns = ctx.primary[*comp_id].continuous_assignments();
+            if !cont_assigns.is_empty() {
+                env.pc.push_continuous_assigns(comp, cont_assigns);
+            }
+
+            // first layout the signature
+            for sig_port in comp_aux.signature().iter() {
+                let def_idx = comp_aux.port_offset_map[sig_port];
+                let info = &ctx.secondary[def_idx];
+                let idx = env.ports.push(PortValue::new_undef());
+
+                // the direction attached to the port is reversed for the signature.
+                // We only want to add the input ports to the control ports list.
+                if !info.is_data && info.direction != calyx_ir::Direction::Input
+                {
+                    env.control_ports
+                        .insert(idx, info.width.try_into().unwrap());
+                }
+                debug_assert_eq!(index_bases + sig_port, idx);
+            }
+            // second group ports
+            for group_idx in comp_aux.definitions.groups() {
+                //go
+                let go = env.ports.push(PortValue::new_undef());
+                //done
+                let done = env.ports.push(PortValue::new_undef());
+                // quick sanity check asserts
+                let go_actual = index_bases + ctx.primary[group_idx].go;
+                let done_actual = index_bases + ctx.primary[group_idx].done;
+                // Case 1 - Go defined before done
+                if ctx.primary[group_idx].go < ctx.primary[group_idx].done {
+                    debug_assert_eq!(done, done_actual);
+                    debug_assert_eq!(go, go_actual);
+                }
+                // Case 2 - Done defined before go
+                else {
+                    // in this case go is defined after done, so our variable names
+                    // are backward, but this is not a problem since they are
+                    // initialized to the same value
+                    debug_assert_eq!(go, done_actual);
+                    debug_assert_eq!(done, go_actual);
+                }
+                env.control_ports.insert(go, 1);
+                env.control_ports.insert(done, 1);
+            }
+
+            // ref cells and ports are initialized to None
+            for (ref_cell, def_idx) in comp_aux.ref_cell_offset_map.iter() {
+                let info = &ctx.secondary[*def_idx];
+
+                for port_idx in info.ports.iter() {
+                    let port_actual = env.ref_ports.push(None);
+                    debug_assert_eq!(
+                        &env.cells[comp].as_comp().unwrap().index_bases
+                            + port_idx,
+                        port_actual
+                    );
+                }
+                let cell_actual = env.ref_cells.push(None);
+                debug_assert_eq!(
+                    &env.cells[comp].as_comp().unwrap().index_bases + ref_cell,
+                    cell_actual
+                )
+            }
+
+            for (cell_off, def_idx) in comp_aux.cell_offset_map.iter() {
+                let info = &ctx.secondary[*def_idx];
+                if !info.prototype.is_component() {
+                    let port_base = env.ports.peek_next_idx();
+                    for port in info.ports.iter() {
+                        let idx = env.ports.push(PortValue::new_undef());
+                        debug_assert_eq!(
+                            &env.cells[comp].as_comp().unwrap().index_bases
+                                + port,
+                            idx
+                        );
+                        let def_idx = comp_aux.port_offset_map[port];
+                        let port_info = &ctx.secondary[def_idx];
+                        if !(port_info.direction == calyx_ir::Direction::Output
+                            || port_info.is_data && info.is_data)
+                        {
+                            env.control_ports.insert(
+                                idx,
+                                port_info.width.try_into().unwrap(),
+                            );
+                        }
+                    }
+                    let cell_dyn = primitives::build_primitive(
+                        info,
+                        port_base,
+                        env.cells.peek_next_idx(),
+                        ctx,
+                        data_map,
+                        memories_initialized,
+                        check_race.then_some(&mut env.clocks),
+                    );
+                    let cell = env.cells.push(cell_dyn);
+
+                    debug_assert_eq!(
+                        &env.cells[comp].as_comp().unwrap().index_bases
+                            + cell_off,
+                        cell
+                    );
+                } else {
+                    let child_comp = info.prototype.as_component().unwrap();
+                    let child_comp = CellLedger::new_comp(*child_comp, env);
+
+                    let cell = env.cells.push(child_comp);
+                    debug_assert_eq!(
+                        &env.cells[comp].as_comp().unwrap().index_bases
+                            + cell_off,
+                        cell
+                    );
+
+                    // layout sub-component but don't include the data map
+                    layout_component(
+                        env,
+                        ctx,
+                        cell,
+                        &None,
+                        memories_initialized,
+                        check_race,
+                    );
+                }
+            }
+
+            if let Some(data) = data_map {
+                for dec in data.header.memories.iter() {
+                    if !memories_initialized.contains(&dec.name) {
+                        // TODO griffin: maybe make this an error?
+                        warn!(
+                            env.logger,
+                            "Initialization was provided for memory {} but no such memory exists in the entrypoint component.",
+                            dec.name
+                        );
+                    }
+                }
+            }
+        }
+
+        let root = ctx.as_ref().entry_point;
+        let aux = &ctx.as_ref().secondary[root];
+
+        let mut clocks = ClockMap::new();
+        let root_clock = clocks.new_clock();
+        let continuous_clock = clocks.new_clock();
+
+        let mut env = Self {
+            ports: PortMap::with_capacity(aux.port_offset_map.count()),
+            cells: CellMap::with_capacity(aux.cell_offset_map.count()),
+            ref_cells: RefCellMap::with_capacity(
+                aux.ref_cell_offset_map.count(),
+            ),
+            ref_ports: RefPortMap::with_capacity(
+                aux.ref_port_offset_map.count(),
+            ),
+            pc: ProgramCounter::new_empty(),
+            ctx: ctx.clone(),
+            clocks,
+            thread_map: ThreadMap::new(root_clock, continuous_clock),
+            memory_header: None,
+            control_ports: FxHashMap::new(),
+            logger: initialize_logger(logging_conf),
+            ports_to_cells_map: SecondaryMap::new_with_default(0.into()),
+            pinned_ports: PinnedPorts::new(),
+        };
+
+        let ctx = ctx.as_ref();
+
+        let root_node = CellLedger::new_comp(root, &env);
+        let root_cell = env.cells.push(root_node);
+        layout_component(
+            &mut env,
+            ctx,
+            root_cell,
+            &data_map,
+            &mut HashSet::new(),
+            check_race,
+        );
+
+        let root_thread = ThreadMap::root_thread();
+        env.clocks[root_clock].increment(&root_thread);
+        env.clocks[continuous_clock].increment(&ThreadMap::continuous_thread());
+
+        // Initialize program counter
+        // TODO griffin: Maybe refactor into a separate function
+        for (idx, ledger) in env.cells.iter() {
+            if let CellLedger::Component(comp) = ledger {
+                let comp_info = &ctx.primary[comp.comp_id];
+                if !comp_info.is_comb() {
+                    if let Some(ctrl) = comp_info.as_standard().unwrap().control
+                    {
+                        env.pc.vec_mut().push((
+                            if comp.comp_id == root {
+                                Some(root_thread)
+                            } else {
+                                None
+                            },
+                            ControlPoint {
+                                comp: idx,
+                                control_node_idx: ctrl,
+                            },
+                        ))
+                    }
+                }
+            }
+        }
+
+        if let Some(header) = data_map {
+            env.memory_header = Some(header.header.memories);
+        }
+
+        env.ports_to_cells_map =
+            SecondaryMap::capacity_with_default(0.into(), env.ports.len());
+
+        for (cell, ledger) in env.cells.iter() {
+            if let Some(comp_ledger) = ledger.as_comp() {
+                for port in comp_ledger.signature_ports(ctx) {
+                    env.ports_to_cells_map.insert(port, cell);
+                }
+            } else {
+                let cell_dyn = ledger.as_primitive().unwrap();
+                for port in cell_dyn.get_ports().iter_all() {
+                    env.ports_to_cells_map.insert(port, cell);
+                }
+            }
+        }
+        env
+    }
+
     /// A utility function to get the context from the environment. This is not
     /// suitable for cases in which mutation is required. In such cases, the
     /// context should be accessed directly.
@@ -490,273 +424,6 @@ impl<C: AsRef<Context> + Clone> Environment<C> {
                     })
                     .collect_vec();
                 info
-            }
-        }
-    }
-
-    pub fn new(
-        ctx: C,
-        data_map: Option<DataDump>,
-        check_race: bool,
-        logging_conf: LoggingConfig,
-    ) -> Self {
-        let root = ctx.as_ref().entry_point;
-        let aux = &ctx.as_ref().secondary[root];
-
-        let mut clocks = ClockMap::new();
-        let root_clock = clocks.new_clock();
-        let continuous_clock = clocks.new_clock();
-
-        let mut env = Self {
-            ports: PortMap::with_capacity(aux.port_offset_map.count()),
-            cells: CellMap::with_capacity(aux.cell_offset_map.count()),
-            ref_cells: RefCellMap::with_capacity(
-                aux.ref_cell_offset_map.count(),
-            ),
-            ref_ports: RefPortMap::with_capacity(
-                aux.ref_port_offset_map.count(),
-            ),
-            pc: ProgramCounter::new_empty(),
-            clocks,
-            thread_map: ThreadMap::new(root_clock, continuous_clock),
-            ctx,
-            memory_header: None,
-            pinned_ports: PinnedPorts::new(),
-            control_ports: FxHashMap::new(),
-            logger: logging::initialize_logger(logging_conf),
-            group_instances: vec![],
-            ports_to_cells_map: SecondaryMap::new_with_default(0.into()),
-        };
-
-        let root_node = CellLedger::new_comp(root, &env);
-        let root_cell = env.cells.push(root_node);
-        env.layout_component(
-            root_cell,
-            &data_map,
-            &mut HashSet::new(),
-            check_race,
-        );
-
-        let root_thread = ThreadMap::root_thread();
-        env.clocks[root_clock].increment(&root_thread);
-        env.clocks[continuous_clock].increment(&ThreadMap::continuous_thread());
-
-        // Initialize program counter
-        // TODO griffin: Maybe refactor into a separate function
-        for (idx, ledger) in env.cells.iter() {
-            if let CellLedger::Component(comp) = ledger {
-                let comp_info = &env.ctx.as_ref().primary[comp.comp_id];
-                if !comp_info.is_comb() {
-                    if let Some(ctrl) = comp_info.as_standard().unwrap().control
-                    {
-                        env.pc.vec_mut().push((
-                            if comp.comp_id == root {
-                                Some(root_thread)
-                            } else {
-                                None
-                            },
-                            ControlPoint {
-                                comp: idx,
-                                control_node_idx: ctrl,
-                            },
-                        ))
-                    }
-                }
-            }
-        }
-
-        if let Some(header) = data_map {
-            env.memory_header = Some(header.header.memories);
-        }
-
-        env.ports_to_cells_map =
-            SecondaryMap::capacity_with_default(0.into(), env.ports.len());
-
-        for (cell, ledger) in env.cells.iter() {
-            if let Some(comp_ledger) = ledger.as_comp() {
-                let ctx = env.ctx.as_ref();
-                for group in
-                    ctx.secondary[comp_ledger.comp_id].definitions.groups()
-                {
-                    env.group_instances.push((cell, group))
-                }
-
-                for port in comp_ledger.signature_ports(ctx) {
-                    env.ports_to_cells_map.insert(port, cell);
-                }
-            } else {
-                let cell_dyn = ledger.as_primitive().unwrap();
-                for port in cell_dyn.get_ports().iter_all() {
-                    env.ports_to_cells_map.insert(port, cell);
-                }
-            }
-        }
-
-        env
-    }
-
-    /// Internal function used to layout a given component from a cell id
-    ///
-    /// Layout is handled in the following order:
-    /// 1. component signature (input/output)
-    /// 2. group hole ports
-    /// 3. cells + ports, primitive
-    /// 4. sub-components
-    /// 5. ref-cells & ports
-    fn layout_component(
-        &mut self,
-        comp: GlobalCellIdx,
-        data_map: &Option<DataDump>,
-        memories_initialized: &mut HashSet<String>,
-        check_race: bool,
-    ) {
-        // for mutability reasons, see note in `[Environment::new]`
-        let ctx = self.ctx.clone();
-        let ctx_ref = ctx.as_ref();
-
-        let ComponentLedger {
-            index_bases,
-            comp_id,
-        } = self.cells[comp]
-            .as_comp()
-            .expect("Called layout component with a non-component cell.");
-        let comp_aux = &ctx_ref.secondary[*comp_id];
-
-        // Insert the component's continuous assignments into the program counter, if non-empty
-        let cont_assigns =
-            self.ctx.as_ref().primary[*comp_id].continuous_assignments();
-        if !cont_assigns.is_empty() {
-            self.pc.push_continuous_assigns(comp, cont_assigns);
-        }
-
-        // first layout the signature
-        for sig_port in comp_aux.signature().iter() {
-            let def_idx = comp_aux.port_offset_map[sig_port];
-            let info = &self.ctx.as_ref().secondary[def_idx];
-            let idx = self.ports.push(PortValue::new_undef());
-
-            // the direction attached to the port is reversed for the signature.
-            // We only want to add the input ports to the control ports list.
-            if !info.is_data && info.direction != calyx_ir::Direction::Input {
-                self.control_ports
-                    .insert(idx, info.width.try_into().unwrap());
-            }
-            debug_assert_eq!(index_bases + sig_port, idx);
-        }
-        // second group ports
-        for group_idx in comp_aux.definitions.groups() {
-            //go
-            let go = self.ports.push(PortValue::new_undef());
-
-            //done
-            let done = self.ports.push(PortValue::new_undef());
-
-            // quick sanity check asserts
-            let go_actual =
-                index_bases + self.ctx.as_ref().primary[group_idx].go;
-            let done_actual =
-                index_bases + self.ctx.as_ref().primary[group_idx].done;
-            // Case 1 - Go defined before done
-            if self.ctx.as_ref().primary[group_idx].go
-                < self.ctx.as_ref().primary[group_idx].done
-            {
-                debug_assert_eq!(done, done_actual);
-                debug_assert_eq!(go, go_actual);
-            }
-            // Case 2 - Done defined before go
-            else {
-                // in this case go is defined after done, so our variable names
-                // are backward, but this is not a problem since they are
-                // initialized to the same value
-                debug_assert_eq!(go, done_actual);
-                debug_assert_eq!(done, go_actual);
-            }
-            self.control_ports.insert(go, 1);
-            self.control_ports.insert(done, 1);
-        }
-
-        // ref cells and ports are initialized to None
-        for (ref_cell, def_idx) in comp_aux.ref_cell_offset_map.iter() {
-            let info = &self.ctx.as_ref().secondary[*def_idx];
-
-            for port_idx in info.ports.iter() {
-                let port_actual = self.ref_ports.push(None);
-                debug_assert_eq!(
-                    &self.cells[comp].as_comp().unwrap().index_bases + port_idx,
-                    port_actual
-                );
-            }
-            let cell_actual = self.ref_cells.push(None);
-            debug_assert_eq!(
-                &self.cells[comp].as_comp().unwrap().index_bases + ref_cell,
-                cell_actual
-            )
-        }
-
-        for (cell_off, def_idx) in comp_aux.cell_offset_map.iter() {
-            let info = &self.ctx.as_ref().secondary[*def_idx];
-            if !info.prototype.is_component() {
-                let port_base = self.ports.peek_next_idx();
-                for port in info.ports.iter() {
-                    let idx = self.ports.push(PortValue::new_undef());
-                    debug_assert_eq!(
-                        &self.cells[comp].as_comp().unwrap().index_bases + port,
-                        idx
-                    );
-                    let def_idx = comp_aux.port_offset_map[port];
-                    let port_info = &self.ctx.as_ref().secondary[def_idx];
-                    if !(port_info.direction == calyx_ir::Direction::Output
-                        || port_info.is_data && info.is_data)
-                    {
-                        self.control_ports
-                            .insert(idx, port_info.width.try_into().unwrap());
-                    }
-                }
-                let cell_dyn = primitives::build_primitive(
-                    info,
-                    port_base,
-                    self.cells.peek_next_idx(),
-                    self.ctx.as_ref(),
-                    data_map,
-                    memories_initialized,
-                    check_race.then_some(&mut self.clocks),
-                );
-                let cell = self.cells.push(cell_dyn);
-
-                debug_assert_eq!(
-                    &self.cells[comp].as_comp().unwrap().index_bases + cell_off,
-                    cell
-                );
-            } else {
-                let child_comp = info.prototype.as_component().unwrap();
-                let child_comp = CellLedger::new_comp(*child_comp, self);
-
-                let cell = self.cells.push(child_comp);
-                debug_assert_eq!(
-                    &self.cells[comp].as_comp().unwrap().index_bases + cell_off,
-                    cell
-                );
-
-                // layout sub-component but don't include the data map
-                self.layout_component(
-                    cell,
-                    &None,
-                    memories_initialized,
-                    check_race,
-                );
-            }
-        }
-
-        if let Some(data) = data_map {
-            for dec in data.header.memories.iter() {
-                if !memories_initialized.contains(&dec.name) {
-                    // TODO griffin: maybe make this an error?
-                    warn!(
-                        self.logger,
-                        "Initialization was provided for memory {} but no such memory exists in the entrypoint component.",
-                        dec.name
-                    );
-                }
             }
         }
     }
@@ -2971,7 +2638,7 @@ impl<C: AsRef<Context> + Clone> BaseSimulator<C> {
                                     }
                                 };
 
-                                if changed.as_bool() {
+                                if changed.as_bool() && !rerun_all_primitives {
                                     changed_cells.insert(
                                         self.env.ports_to_cells_map[dest],
                                     );
