+++ cider/src/flatten/structures/environment/mod.rs
@@ -1,11 +1,12 @@
 mod assignments;
 pub mod clock;
 mod env;
+mod maps;
 mod program_counter;
 mod traverser;
 mod wave;
 
-pub use env::{BaseSimulator, Environment, PortMap, Simulator};
+pub use env::{BaseSimulator, Environment, Simulator};
+pub(crate) use maps::CellLedger;
+pub use maps::PortMap;
 pub use traverser::{Path, PathError, PathResolution};
-
-pub(crate) use env::CellLedger;
